♻️ :: UserApi -> AuthApi 리팩토링
